# perf: rework scalar filter kernels