--- a/crates/polars-compute/src/filter/mod.rs
+++ b/crates/polars-compute/src/filter/mod.rs
@@ -4,32 +4,14 @@ mod primitive;
 mod scalar;
 
 use arrow::array::growable::make_growable;
-use arrow::array::*;
-use arrow::bitmap::utils::{BitChunkIterExact, BitChunksExact, SlicesIterator};
-use arrow::bitmap::{Bitmap, MutableBitmap};
+use arrow::array::{new_empty_array, Array, BinaryViewArray, BooleanArray, PrimitiveArray};
+use arrow::bitmap::utils::SlicesIterator;
 use arrow::datatypes::ArrowDataType;
-use arrow::types::simd::Simd;
-use arrow::types::{BitChunkOnes, NativeType};
 use arrow::with_match_primitive_type_full;
-use boolean::*;
-use polars_error::*;
-use primitive::*;
-
-/// Function that can filter arbitrary arrays
-pub type Filter<'a> = Box<dyn Fn(&dyn Array) -> Box<dyn Array> + 'a + Send + Sync>;
-
-#[inline]
-fn get_leading_ones(chunk: u64) -> u32 {
-    if cfg!(target_endian = "little") {
-        chunk.trailing_ones()
-    } else {
-        chunk.leading_ones()
-    }
-}
+use polars_error::PolarsResult;
 
 pub fn filter(array: &dyn Array, mask: &BooleanArray) -> PolarsResult<Box<dyn Array>> {
-    // The validities may be masking out `true` bits, making the filter operation
-    // based on the values incorrect
+    // Treat null mask values as false.
     if let Some(validities) = mask.validity() {
         let values = mask.values();
         let new_values = values & validities;
@@ -50,29 +32,31 @@ pub fn filter(array: &dyn Array, mask: &BooleanArray) -> PolarsResult<Box<dyn Ar
     use arrow::datatypes::PhysicalType::*;
     match array.data_type().to_physical_type() {
         Primitive(primitive) => with_match_primitive_type_full!(primitive, |$T| {
-            let array = array.as_any().downcast_ref().unwrap();
-            // Ok(Box::new(filter_primitive::<$T>(array, mask.values())))
-            Ok(Box::new(scalar::filter_primitive_scalar::<$T>(array, mask.values())))
+            let array: &PrimitiveArray<$T> = array.as_any().downcast_ref().unwrap();
+            let (values, validity) = primitive::filter_values_and_validity::<$T>(array.values(), array.validity(), mask.values());
+            Ok(Box::new(PrimitiveArray::from_vec(values).with_validity(validity)))
         }),
         Boolean => {
             let array = array.as_any().downcast_ref::<BooleanArray>().unwrap();
-            let (values, validity) =
-                filter_bitmap_and_validity(array.values(), array.validity(), mask.values());
+            let (values, validity) = boolean::filter_bitmap_and_validity(
+                array.values(),
+                array.validity(),
+                mask.values(),
+            );
             Ok(BooleanArray::new(array.data_type().clone(), values, validity).boxed())
         },
         BinaryView => {
             let array = array.as_any().downcast_ref::<BinaryViewArray>().unwrap();
             let views = array.views();
             let validity = array.validity();
-            // TODO! we might opt for a filter that maintains the bytes_count
-            // currently we don't do that and bytes_len is set to UNKNOWN.
-            let (views, validity) = filter_values_and_validity(views, validity, mask.values());
+            let (views, validity) =
+                primitive::filter_values_and_validity(views, validity, mask.values());
             Ok(unsafe {
                 BinaryViewArray::new_unchecked_unknown_md(
                     array.data_type().clone(),
                     views.into(),
                     array.data_buffers().clone(),
-                    validity.map(|v| v.freeze()),
+                    validity.map(|v| v),
                     Some(array.total_buffer_len()),
                 )
             }