This PR uses the new boolean filters to handle nulls in all our filters, and has a reworked scalar path filter. It also lays the groundwork for easily slotting in specific SIMD paths instead of the scalar later.

Modest speedups all-around, with only very specific combinations being slightly slower. Below are some sample benchmarks for n = 10^8 on Apple M1.

Without any nulls in the payload:

```
selectivity       main         branch  speedup
u8
0.00001     0.03067962     0.00971654     3.2x
0.00010     0.03578900     0.01379362     2.6x
0.00100     0.05605988     0.02842633     2.0x
0.01000     0.10832329     0.05732238     1.9x
0.10000     0.18866021     0.14063879     1.3x
0.50000     0.38846837     0.25560483     1.5x
0.90000     0.62037029     0.28825442     2.2x
0.99000     0.41020429     0.22628954     1.8x
0.99900     0.15035492     0.12312429     1.2x
0.99990     0.12071196     0.11451621     1.1x
0.99999     0.11959000     0.11468092     1.0x
f64
0.00001     0.02904350     0.01133346     2.6x
0.00010     0.03190525     0.01483838     2.2x
0.00100     0.05377196     0.04013158     1.3x
0.01000     0.16904179     0.12135558     1.4x
0.10000     0.27002258     0.23378392     1.2x
0.50000     0.69589496     0.56552742     1.2x
0.90000     1.09929317     0.81537075     1.3x
0.99000     1.05055963     0.86166412     1.2x
0.99900     0.90991829     0.86368392     1.1x
0.99990     0.84915117     0.90556942     0.9x
0.99999     0.89902233     0.86255267     1.0x
str
0.00001     0.03019371     0.01210613     2.5x
0.00010     0.03551521     0.01880279     1.9x
0.00100     0.08062025     0.06540263     1.2x
0.01000     0.18216504     0.13414883     1.4x
0.10000     0.42839887     0.43168421     1.0x
0.50000     1.03755975     0.99724696     1.0x
0.90000     1.75257592     1.56434954     1.1x
0.99000     1.81074688     1.70101721     1.1x
0.99900     1.75836221     1.71458492     1.0x
0.99990     1.78606162     1.79906375     1.0x
0.99999     1.75107996     1.70657762     1.0x
```

With 10% of the payload being nulls:

```
selectivity       main         branch  speedup
u8
0.00001     0.04466854     0.01629858     2.7x
0.00010     0.05015971     0.02353896     2.1x
0.00100     0.08562458     0.04379925     2.0x
0.01000     0.14814654     0.15693012     0.9x
0.10000     0.31246542     0.30595038     1.0x
0.50000     1.15083700     0.35854967     3.2x
0.90000     2.06214100     0.39735929     5.2x
0.99000     1.23677525     0.35589008     3.5x
0.99900     0.38117267     0.19885325     1.9x
0.99990     0.26254242     0.17230442     1.5x
0.99999     0.25265775     0.17859104     1.4x
f64
0.00001     0.04572262     0.01910842     2.4x
0.00010     0.05307171     0.02670975     2.0x
0.00100     0.09755967     0.06618713     1.5x
0.01000     0.29210800     0.23889150     1.2x
0.10000     0.41639846     0.42964929     1.0x
0.50000     1.50237033     0.69708017     2.2x
0.90000     2.65734029     0.99554375     2.7x
0.99000     1.92291908     1.07280654     1.8x
0.99900     1.12154263     1.01173225     1.1x
0.99990     0.99597621     0.99489346     1.0x
0.99999     0.98589592     0.98362196     1.0x
str
0.00001     0.04470687     0.01981188     2.3x
0.00010     0.05394471     0.02838900     1.9x
0.00100     0.17123421     0.07548829     2.3x
0.01000     0.31859496     0.25365346     1.3x
0.10000     0.61130029     0.67295100     0.9x
0.50000     1.84593263     1.20410746     1.5x
0.90000     3.31969500     1.80379200     1.8x
0.99000     2.75647108     1.98157950     1.4x
0.99900     2.03973321     1.95453050     1.0x
0.99990     1.97040792     1.94661246     1.0x
0.99999     1.92459388     1.92393138     1.0x
```


Benchmark code:
```python
import polars as pl
import numpy as np
from timeit import timeit

n = 10**8
selectivities = [0, 0.00001, 0.0001, 0.001, 0.01, 0.1, 0.5, 0.9, 0.99, 0.999, 0.9999, 0.99999, 1]

rng = np.random.Generator(np.random.PCG64(42))
base_payload = pl.Series(rng.uniform(size = n))

types = [
    ("u8", (base_payload * (256 - 1e-6)).cast(pl.UInt8)),
    ("f64", base_payload),
    ("str", base_payload.cast(pl.String)),
]

for typ, payload in types:
  print(f"polars timings {typ}")
  for selectivity in selectivities:
      threshold = selectivity if selectivity < 1 else 1.1  # Just to be safe.
      filt = rng.uniform(size = n) < threshold
      df = pl.DataFrame({"filt": filt, "payload": payload})
      time = timeit(lambda: df.select(pl.col.payload.filter(pl.col.filt)), number=10)
      print(f"{selectivity:.5f}  {time:11.8f}")

  print(f"polars timings {typ} with 10% nulls")
  for selectivity in selectivities:
      threshold = selectivity if selectivity < 1 else 1.1  # Just to be safe.
      filt = rng.uniform(size = n) < threshold
      is_not_null = rng.uniform(size = n) < 0.9
      df = pl.DataFrame({"filt": filt, "payload": payload, "is_not_null": is_not_null})
      df = df.select(pl.col.filt, pl.when(pl.col.is_not_null).then(pl.col.payload))
      time = timeit(lambda: df.select(pl.col.payload.filter(pl.col.filt)), number=10)
      print(f"{selectivity:.5f}  {time:11.8f}")
```